Edit Action Post and Project Delete Functionality
Edited action post function so that posted actions are associated to the project that it is posted under rather than becoming a free agent.

Edited project delete function so that all associated actions are also removed when linked project is removed. Added removeByProjectGroup function to db file according to follow up lecture that explained how to fix associative delete functionality.